requirements-dev.txt
annotated-types==0.7.0
asgiref==3.9.1
boto3==1.39.10
botocore==1.39.10
Django==5.2.4
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-phonenumber-field==8.1.0
djangorestframework-api-key==3.1.0
gunicorn==23.0.0
injector==0.22.0
jmespath==1.0.1
packaging==25.0
phonenumberslite==9.0.10
pillow==11.3.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
whitenoise==6.9.0
astroid==3.3.11
black==25.1.0
click==8.2.1
contextlib2==21.6.0
dill==0.4.0
django-debug-toolbar==5.2.0
django-extensions==4.1
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
pathspec==0.12.1
platformdirs==4.3.8
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
tomlkit==0.13.3